+++ src/mapping.hpp
@@ -10,44 +10,11 @@
 #include <igl/edges.h>
 #include <algorithm>
 #include <Eigen/Core>
+#include <clipper2/clipper.h>
+#include <igl/boundary_loop.h>
 
 /**
- * Helper function to extract triangle vertices at a specific face index.
- *
- * @param F_ The face matrix of the triangle mesh.
- * @param faceID Index of the face to extract.
- * @param V The vertex matrix.
- * @param A Output parameter for the first vertex of the triangle.
- * @param B Output parameter for the second vertex of the triangle.
- * @param C Output parameter for the third vertex of the triangle.
- */
-void get_triface(
-    const compas::RowMatrixXi& F_, 
-    int faceID, 
-    const compas::RowMatrixXd& V, 
-    compas::RowMatrixXd& A, 
-    compas::RowMatrixXd& B, 
-    compas::RowMatrixXd& C);
-
-/**
- * Map a point to a mesh using barycentric coordinates.
- *
- * @param F_ The face matrix of the triangle mesh.
- * @param UV_ The UV coordinates matrix.
- * @param pt The 3D point to map.
- * @param barycentric Output parameter for the barycentric coordinates.
- * @param faceID Output parameter for the face index.
- * @return True if the point was successfully mapped, false otherwise.
- */
-bool map_point3d_simple(
-    const compas::RowMatrixXi& F_, 
-    const compas::RowMatrixXd& UV_, 
-    const Eigen::Vector3d& pt, 
-    Eigen::Vector3d& barycentric, 
-    int& faceID);
-
-/**
- * Map a 2D pattern mesh onto a 3D target mesh using AABB tree-based mapping.
+ * Map a 2D already croppedpattern mesh onto a 3D target mesh using AABB tree-based mapping.
  * This function maps each vertex of the pattern mesh to the target mesh using
  * barycentric interpolation based on the UV parameterization of the target mesh.
  *
@@ -59,7 +26,7 @@ bool map_point3d_simple(
  * @param pattern_uv The UV coordinates of the pattern mesh.
  * @return Vector of vectors representing the polygonal faces of the mapped pattern mesh.
  */
-std::vector<std::vector<int>> map_mesh(
+std::vector<std::vector<int>> map_mesh_cropped(
     Eigen::Ref<const compas::RowMatrixXd> v, 
     Eigen::Ref<const compas::RowMatrixXi> f, 
     Eigen::Ref<const compas::RowMatrixXd> uv,
@@ -67,6 +34,46 @@ std::vector<std::vector<int>> map_mesh(
     const std::vector<std::vector<int>>& pattern_f  , 
     Eigen::Ref<const compas::RowMatrixXd> pattern_uv);
 
+/**
+ * Check if two paths intersect.
+ *
+ * @param path1 The first path.
+ * @param path2 The second path.
+ * @param scale The scale factor for the paths.
+ * @return True if the paths intersect, false otherwise.
+ */
+bool path_intersect(const Clipper2Lib::PathD& path1, const Clipper2Lib::PathD& path2, double scale = 1e6);
+
+/**
+ * Check if two paths intersect.
+ *
+ * @param paths1 The first path.
+ * @param paths2 The second path.
+ * @param scale The scale factor for the paths.
+ * @return True if the paths intersect, false otherwise.
+ */
+bool paths_intersect(const Clipper2Lib::PathsD& paths1, const Clipper2Lib::PathsD& paths2, double scale = 1e6);
+
+/**
+ * Convert Eigen mesh to Clipper paths.
+ *
+ * @param flattned_target_uv The flattened target UV coordinates.
+ * @param target_f The target mesh face indices.
+ * @param pattern_v The pattern mesh vertex coordinates.
+ * @param pattern_f The pattern mesh face indices.
+ * @param clip_boundaries Whether to clip the pattern mesh to the boundaries of the target mesh.
+ * @param tolerance The tolerance for point comparison, to remove duplicates.
+ * @return A tuple of the clipped pattern mesh vertex coordinates and face indices.
+ */
+std::tuple<compas::RowMatrixXd, std::vector<std::vector<int>>> eigen_to_clipper (
+    Eigen::Ref<const compas::RowMatrixXd> flattned_target_uv,
+    Eigen::Ref<const compas::RowMatrixXi> target_f, 
+    
+    Eigen::Ref<const compas::RowMatrixXd> pattern_v, 
+    const std::vector<std::vector<int>>& pattern_f,
+    bool clip_boundaries,
+    double tolerance = 1e-6
+)  ;    
 
 /**
  * Map a 2D pattern mesh onto a 3D target mesh with automatic parameterization, mesh is always mapped to 1x1 unit.
@@ -77,10 +84,14 @@ std::vector<std::vector<int>> map_mesh(
  * @param target_f #F x 3 matrix of target mesh triangle indices
  * @param pattern_v #V x 3 matrix of pattern mesh vertex coordinates
  * @param pattern_f vector of vectors of int of pattern mesh triangle indices
- * @return A vector of face indices representing polygons in the mapped mesh
+ * @param clip_boundaries whether to clip the pattern mesh to the boundaries of the target mesh
+ * @return A tuple of pattern_v and face indices representing polygons in the mapped mesh
  */
-std::vector<std::vector<int>> map_mesh_with_automatic_parameterization(
+std::tuple<compas::RowMatrixXd, std::vector<std::vector<int>>> map_mesh_with_automatic_parameterization(
     Eigen::Ref<const compas::RowMatrixXd> target_v, 
     Eigen::Ref<const compas::RowMatrixXi> target_f, 
     Eigen::Ref<compas::RowMatrixXd> pattern_v, 
-    const std::vector<std::vector<int>>& pattern_f);
+    const std::vector<std::vector<int>>& pattern_f,
+    bool clip_boundaries,
+    double tolerance = 1e-6
+);