+++ tests/test_boundaries.py
@@ -1,5 +1,5 @@
 import compas
-import compas_libigl
+from compas_libigl.boundaries import trimesh_boundaries
 from compas.datastructures import Mesh
 
 
@@ -7,5 +7,5 @@ def test_trimesh_boundaries():
     mesh = Mesh.from_off(compas.get("tubemesh.off"))
     mesh.quads_to_triangles()
     M = mesh.to_vertices_and_faces()
-    boundaries = compas_libigl.trimesh_boundaries(M)
+    boundaries = trimesh_boundaries(M)
     assert len(boundaries) == 1