+++ tests/test_geodistance.py
@@ -1,5 +1,5 @@
 import compas
-import compas_libigl as libigl
+from compas_libigl.geodistance import trimesh_geodistance, trimesh_geodistance_multiple
 from compas.datastructures import Mesh
 
 
@@ -8,9 +8,9 @@ def test_trimesh_geodistance():
     mesh.quads_to_triangles()
     M = mesh.to_vertices_and_faces()
     source = 0
-    distances = libigl.trimesh_geodistance(M, source, method="exact")
+    distances = trimesh_geodistance(M, source, method="exact")
     assert len(distances) == mesh.number_of_vertices()
-    distances = libigl.trimesh_geodistance(M, source, method="heat")
+    distances = trimesh_geodistance(M, source, method="heat")
     assert len(distances) == mesh.number_of_vertices()
 
 
@@ -19,7 +19,7 @@ def test_trimesh_geodistance_multiple():
     mesh.quads_to_triangles()
     M = mesh.to_vertices_and_faces()
     sources = [0, 1]
-    distances = libigl.trimesh_geodistance_multiple(M, sources, method="exact")
+    distances = trimesh_geodistance_multiple(M, sources, method="exact")
     assert len(distances) == mesh.number_of_vertices()
-    distances = libigl.trimesh_geodistance_multiple(M, sources, method="heat")
+    distances = trimesh_geodistance_multiple(M, sources, method="heat")
     assert len(distances) == mesh.number_of_vertices()