+++ tests/test_intersections.py
@@ -1,5 +1,5 @@
 import compas
-import compas_libigl
+from compas_libigl.intersections import intersection_ray_mesh, intersection_rays_mesh
 from compas.datastructures import Mesh
 from compas.geometry import add_vectors, scale_vector
 
@@ -10,7 +10,7 @@ def test_intersection_ray_mesh():
     M = mesh.to_vertices_and_faces()
     centroid = mesh.centroid()
     ray = [centroid[0], centroid[1], 0], [0, 0, 1.0]
-    hits = compas_libigl.intersection_ray_mesh(ray, M)
+    hits = intersection_ray_mesh(ray, M)
     assert len(hits) == 1
     # Test computing intersection point
     point = add_vectors(ray[0], scale_vector(ray[1], hits[0][3]))
@@ -23,5 +23,5 @@ def test_intersection_rays_mesh():
     M = mesh.to_vertices_and_faces()
     centroid = mesh.centroid()
     rays = [([centroid[0], centroid[1], 0], [0, 0, 1.0]), ([centroid[0], centroid[1], 1], [0, 0, -1.0])]
-    hits = compas_libigl.intersection_rays_mesh(rays, M)
+    hits = intersection_rays_mesh(rays, M)
     assert len(hits) == 2